Add alternative test for user name input
This is a better test since it tests the actual value typed is equal to
"test" and not just a non-empty string.

This actually failed when using Xcode 13 beta as the text actually
typed was only "tst". Not sure why - could be a bug in the beta
software, or it could be a timing issue.